Refactor routers/Add availability subscription
