+++ services/assistant/app/adapters/input/bot/routers/__init__.py
@@ -1,0 +1,5 @@
+from app.adapters.input.bot.routers.booking import BookingRouter
+from app.adapters.input.bot.routers.proxy import ProxyRouter
+from app.adapters.input.bot.routers.main import MainRouter
+
+__all__ = ["BookingRouter", "MainRouter", "ProxyRouter"]
