+++ services/assistant/app/adapters/input/bot/routers/base.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+from aiogram import Dispatcher
+
+
+class BaseRouter(ABC):
+    """Base class for registering router."""
+
+    @abstractmethod
+    def register_router(self, dp: Dispatcher):
+        pass
